… src/Build/BackEnd/Components/Logging/LoggingServiceLogMethods.cs
… src/Build/BackEnd/Components/RequestBuilder/IntrinsicTasks/ItemGroupLoggingHelper.cs
… src/Build/BackEnd/Components/RequestBuilder/TargetEntry.cs
… src/Build/BackEnd/Components/RequestBuilder/TargetUpToDateChecker.cs
… src/Build/BackEnd/TaskExecutionHost/TaskExecutionHost.cs
… src/Build/Evaluation/Evaluator.cs
… src/Build/Evaluation/PropertyTrackingEvaluatorDataWrapper.cs
… src/Build/Logging/BinaryLogger/BinaryLogger.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsFieldFlags.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsFields.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsReader.cs
… src/Build/Logging/BinaryLogger/BuildEventArgsWriter.cs
… src/Framework/BuildEventArgs.cs
… src/Framework/BuildMessageEventArgs.cs
… src/Framework/BuildStatusEventArgs.cs
… src/Framework/LazyFormattedBuildEventArgs.cs
… src/Framework/ProjectFinishedEventArgs.cs
… src/Framework/ProjectImportedEventArgs.cs
… src/Framework/ProjectStartedEventArgs.cs
… src/Framework/PropertyReassignmentEventArgs.cs
… src/Framework/TargetFinishedEventArgs.cs
… src/Framework/TargetSkippedEventArgs.cs
… src/Framework/TargetStartedEventArgs.cs
… src/Framework/TaskCommandLineEventArgs.cs
… src/Framework/TaskFinishedEventArgs.cs
… src/Framework/TaskParameterEventArgs.cs
… src/Framework/TaskStartedEventArgs.cs
… src/Framework/TelemetryEventArgs.cs
… src/Shared/LogMessagePacketBase.cs
… src/Shared/TaskLoggingHelper.cs