We're almost at the finish line. This PR contains most remaining low-hanging fruit fixes to logging efficiency. There are more fruit of course, but it's no longer clear what else to do to meaningfully reduce binlog size and overhead even further.

The primary insights are:
 * don't log BuildEventArgs.Message where it can be recovered (less strings)
 * log Message separately from Arguments, instead of concatenating into formatted message and logging that (smaller strings, more deduplication)
 * cache string resources instead of retrieving them every time, log resource string and arguments separately
 * add more test coverage for Serial and Parallel console loggers (which is currently sorely lacking, since we mostly use MockLogger in tests)

I tried to separate independent changes into separate commits, so reviewing commit-by-commit is better. For some changes, turn on Ignore Whitespace will help.

Final binlog size and perf numbers are very favorable.

Binlog size:

16.9.2 | main | This PR
-- | -- | --
20 MB | 3.5 MB | 2.7 MB
120 MB | 15 MB | 10 MB

Incremental build duration on a smaller project:

  | 16.9.2 | This PR
-- | -- | --
/m /bl | 18.6 s | 14.1 s
/bl | 32 s | 18 s
/m | 9 s | 9 s
no /m no /bl | 15 s | 15 s

Incremental build for Roslyn.sln:

  | 16.9.2 | This PR
-- | -- | --
/m /bl | 1:11 s | 37 s
/bl | 2:23 s | 1:03 s