@KirillOsenkov 2021-04-05
Fixes https://github.com/dotnet/msbuild/issues/6199