@danmoseley 2021-04-05
> cache string resources instead of retrieving them every time

This is not directly related to /bl right? Is it worth it? You are only saving a hashtable lookup.