@danmoseley 2021-04-05
Some nice savings in bytes across the pipe, I assume this gives measurable improvements for /m /flp:d or similar.