@KirillOsenkov 2021-04-05
@danmoseley I have to confess I didn't know how `GetResourceString` is implemented, so I just assumed there'll be some non-zero amount of overhead there. Now that you're mentioning it, you're right of course, it will likely be negligible even if called thousands of times. I just benchmarked it and even though it's about 5x overhead compared to a simple hashtable lookup, it's still nanoseconds:

https://github.com/KirillOsenkov/Benchmarks/blob/c77e8587477e38a9c7347efa6bb3d08b79f33a02/Benchmarks/Tests/ResourceManager.cs#L13-L15

I was also driven by the fact that messages logged by RAR are among the most numerous and expensive:
![image](https://user-images.githubusercontent.com/679326/113534082-871d5400-9584-11eb-9a22-b416167cea04.png)

So it's likely not worth going after GetResourceString() just for the sake of it, but here it also helps with reducing duplication and preparing indented strings once instead of allocating every time. Allocations are most certainly a problem.