@KirillOsenkov 2021-04-05
Oh, and file loggers are not even worth measuring right now because we know there are too many low hanging fruit there, primarily because of allocations. To give you a rough example:

Build | Duration
-- | --
Single-core no logging: | 15 s
/bl with no TaskParameterEventArgs | 15 s
/bl | 18 s
/flp1:v=diag | 59 s

There's just so much we can fix there. I hope I'll have time this year to do a pass like I did with `/bl`.