@KirillOsenkov 2021-04-05
I just checked the ratio of binlog to diag text log and this is pretty neat:
![image](https://user-images.githubusercontent.com/679326/113541795-c903c580-9597-11eb-9a05-119d7a373d72.png)