@danmoseley 2021-04-05
It would be great to improve /flp1:v=diag but nobody does that normally of course. v=d and v=n would likely be worth more attention..