@KirillOsenkov 2021-04-05
Absolutely, when we look at text logging we'll look at those verbosities first of course.